Move from math to pulse module
